fix(archive): atomic remove-kind path + split test modules
Follow-up to #1555 (merged before these two review fixes landed). Closes
the last two of Wes's review notes.

Fix A: add remove_owner_p_kind / remove_save_subscription_kind

The archive toggle handlers did a TS-side read-modify-overwrite on the
shared owner_p row, and on toggle-OFF called deleteSaveSubscription for
the whole row — silently dropping the *other* kind (24200 vs 44200) if
subs state was stale. Close the loop with a symmetric atomic remove:

- store.rs: add remove_owner_p_kind mirroring merge_owner_p_kinds — same
  BEGIN IMMEDIATE -> closure -> COMMIT/best-effort-ROLLBACK shape; reads
  current kinds, removes the target, DELETEs the row when the list
  becomes empty or UPDATEs to the reduced list otherwise.
- mod.rs: add remove_save_subscription_kind Tauri command delegating to
  store::remove_owner_p_kind; registered in lib.rs alongside the merge cmd.
- tauriArchive.ts: add removeSaveSubscriptionKind wrapper + notify.
- LocalArchiveSettingsCard.tsx: rewrite both toggle handlers to route
  through the atomic commands — toggle-ON merges the kind, toggle-OFF
  removes it. The TS-side read-modify-overwrite and whole-row delete
  branch are gone; subs dropped from both useCallback dep arrays.
- store_tests: 4 unit tests for remove_owner_p_kind.

Fix B: split archive test modules

The check-file-sizes.mjs override for archive/mod.rs had ratcheted
1465->1705 across the PR series. Extract both oversized test blocks into
sibling mod_tests.rs / store_tests.rs files wired via #[path]. mod.rs is
now 532 lines, store.rs 599 — both under 1000. The two mod.rs/store.rs
overrides collapse to a single mod_tests.rs override (1208).
